use int32 rather than long
need less disk space for text data

+++ mkdata/tools/buildata.py
@@ -59,7 +59,7 @@ def handle(srcpf,rspf,srctf,rstf,splf,padv):
 					for i in xrange(tmp):
 						lind=frd.readline().strip().decode("utf-8").split(" ")
 						td.append([int(linu) for linu in lind])
-					td=numpy.array(padmat(td,padv),dtype=long)
+					td=numpy.array(padmat(td,padv),dtype=numpy.int32)
 					rp, td=shufflepair(rp,td)
 					wrtkey=str(cuwid)
 					rpf[wrtkey]=rp
